[AIRFLOW-5786] Move AWS Sns hook and operator No deprecation placeholders yet to preserve original history
